pass all tests
this commit shows that the tests in the test class are not enough to
document what the AnalisadorDeAprovacao class should do. The way the
class is right now passes all the tests, so it the intention of this
exercise is to show how to do TDD, it unfortunately failed =/

the next commit will implement the right intended code (using the grade
criteria).